+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ---
+## [0.6.3b] - 2021-04-25
+### Changed
+- new hexagon background
+- spiced audio visualiser effects
+---
 ## [v0.6.2] - 2021-04-24
 ### Changed
 - Star rating to more closely match classical osu!